# API/cmd/pos-restaurant/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API untuk sistem POS Restoran",
        "title": "POS Restaurant API",
        "contact": {
            "name": "Vincent Ronald",
            "email": "vincentronald9703@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/bills": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bills"
                ],
                "summary": "Ambil semua tagihan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Bill"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bills"
                ],
                "summary": "Buat tagihan untuk sebuah order",
                "parameters": [
                    {
                        "description": "Data tagihan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateBillRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/bills/pay": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bills"
                ],
                "summary": "Proses pembayaran tagihan",
                "parameters": [
                    {
                        "description": "Data pembayaran",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.BillPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/bills/split": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bills"
                ],
                "summary": "Buat tagihan split dari satu order",
                "parameters": [
                    {
                        "description": "Data split bill",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SplitBillRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/bills/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bills"
                ],
                "summary": "Ambil tagihan berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID tagihan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bill"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bills"
                ],
                "summary": "Soft delete tagihan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID tagihan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.BillPaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "bill_id",
                "payment_method"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "bill_id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "reference_number": {
                    "type": "string"
                },
                "room_charge_approved_by": {
                    "type": "integer"
                }
            }
        },
        "handlers.CreateBillRequest": {
            "type": "object",
            "required": [
                "order_id"
            ],
            "properties": {
                "discount_amount": {
                    "type": "number"
                },
                "order_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Bill": {
            "type": "object",
            "properties": {
                "balance_due": {
                    "type": "number"
                },
                "bill_number": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "discount_amount": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "original_bill_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "paid_amount": {
                    "type": "number"
                },
                "service_charge": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "number"
                },
                "tax_amount": {
                    "type": "number"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.SplitBillInput": {
            "type": "object",
            "properties": {
                "discount_amount": {
                    "type": "number"
                },
                "item_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.SplitBillRequest": {
            "type": "object",
            "properties": {
                "original_bill_id": {
                    "description": "\u003c- Tambahan",
                    "type": "integer"
                },
                "original_order_id": {
                    "type": "integer"
                },
                "splits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SplitBillInput"
                    }
                }
            }
        },
...
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    }
}